requirements/develop.txt
astroid==1.6.0
autopep8==1.3.3
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.2.1
Django==2.0.1
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
flake8==3.5.0
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
parso==0.1.1
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.1
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
